blog_backend/requirements.txt
Django==5.0.4
Pillow==10.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0